Consumer Behaviour Insights
This report aims to explore consumer behavior and their attitude towards GenAI in the healthcare industry. The research employs a mixed-methods approach, combining quantitative and qualitative analysis. Data collection was conducted by using online surveys, and all data was analyzed and processed in SPSS. In the analysis, we explore the main effect of independent variable(IV) on dependent variables(DVs:trust and safety). Besides, the study also examines the roles of mediators and moderators. The analysis and results mentioned will be shown in the Finding part. Based on the results of the data collected, we gain insights about the consumer perception of GenAI in healthcare, which provides idea and inspiration for future optimization of AI in healthcare.