Data Science for Marketing
This project aims to provide valuable insights for the Na@onal Tourism Board Organization (NTBO) in a challenging post-pandemic period. Using an approach based on the CRISP-DM model, we divided our analysis into four main notebooks. The first notebook covers Business Understanding, Data Understanding, and Data preparation, laying a solid foundation for the subsequent analyses. In the second notebook, we focus on Market Basket analysis of tourist activities, exploring rule-based patterns of Support, Confidence, and loyalty. The third notebook is dedicated to RFM analysis, focusing mainly on users of segment 144, which is identified as the most promising. Finally, the fourth notebook performs a comparative assessment, analyzing changes in visitor patterns before and after the pandemic and highlighting Portugal's main tourism competitors. Together, these findings aim to help the NTBO better understand visitor patterns and formulate effective strategies to revitalize tourism in the country.